service/src/requirements.txt
Pillow==9.0.1
Appium-Python-Client==2.1.4
aiohttp==3.8.1